requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
faiss-cpu==1.7.4
requests==2.31.0
pydantic==2.5.0
sentence-transformers==2.2.2
torch==2.1.0